add all admin

+++ handlers/statics/admin_panel.py
@@ -3,20 +3,11 @@ from aiogram import types
 from aiogram.dispatcher import FSMContext
 from aiogram.types import CallbackQuery, Message, InlineKeyboardMarkup, InlineKeyboardButton
 from aiogram.dispatcher.filters.state import StatesGroup, State
+
+from handlers.statics.royxatlar import admin_ids, balans_toldirish, balans_ayrish, user_of_banned
 from loader import dp, db, bot
 
-asosiy = [6132434228,343103355]
-admin_ids = [6132434228,343103355]
-admin_qoshish = [6132434228]
-admin_chiqarish = [6132434228]
-balans_toldirish = [6132434228]
-balans_ayrish = [6132434228]
-ban_qilish = []
-bandan_chiqarish = []
-send_message = []
-ruxsatlar = [6132434228]
-delete_admin = []
-user_of_banned = []
+
 class BanStatesGroup(StatesGroup):
     id = State()
     day = State()
@@ -43,7 +34,8 @@ async def admin_panel_commands(message:Message,state:FSMContext):
     msg_4 = f"Kuniga {four[3]} ta qabul qilish, oyiga - > {four[2]} "
     five = await db.select_tarif(tarif_name='fifth')
     msg_5 = f"/start bosilishiga {five[3]} kun bepul qilish "
-    if message.from_user.id in admin_ids:
+    admin = await db.select_admin(telegram_id = message.from_user.id)
+    if admin:
         list_of_commands = (f"/start --> Botni ishga tushirish\n"
                             f"/help --> Yordam\n"
                             f"/statics --> Statistika\n"
@@ -111,12 +103,16 @@ async def balans_id(message:Message,state:FSMContext):
         data = await state.get_data()
         id=data.get("id")
         d = await db.select_user(telegram_id=id)
-        await db.update_balans(telegram_id=d[3],balans=d[7]+int(message.text))
-        d = await db.select_user(telegram_id=id)
-        await message.answer(f"Balans to'ldirildi\nHaydovchi\nusername:{d[1]}\nID:{id}\nBalans:{d[7]} ")
-        await bot.send_message(chat_id=d[3], text=f"Balansingiz {message.text} ga to'ldirildi")
-        await message.delete()
-        await state.finish()
+        if d is not None:
+            await db.update_balans(telegram_id=d[3],balans=d[7]+int(message.text))
+            d = await db.select_user(telegram_id=id)
+            await message.answer(f"Balans to'ldirildi\nHaydovchi\nusername:{d[1]}\nID:{id}\nBalans:{d[7]} ")
+            await bot.send_message(chat_id=d[3], text=f"Balansingiz {message.text} ga to'ldirildi")
+            await message.delete()
+            await state.finish()
+        else:
+            await message.answer("Kechirasiz bunday foydalanuvchi mavjud emas !")
+            await state.finish()
 
 
 @dp.callback_query_handler(text="Balansayrish")
@@ -141,12 +137,17 @@ async def balans_id(message:Message,state:FSMContext):
         data = await state.get_data()
         id=data.get("id")
         d = await db.select_user(telegram_id=id)
-        await db.update_balans(telegram_id=d[3],balans=d[7]+int(message.text))
-        d = await db.select_user(telegram_id=id)
-        await message.answer(f"Balans ayrildi\nHaydovchi\nusername: {d[1]}\nID:{id}\nBalans:{d[7]} ")
-        await bot.send_message(chat_id=d[3], text=f"Balansingiz {message.text} ga kamaytirildi")
-        await message.delete()
-        await state.finish()
+        if d is not None:
+            await db.update_balans(telegram_id=d[3],balans=d[7]+int(message.text))
+            d = await db.select_user(telegram_id=id)
+            await message.answer(f"Balans ayrildi\nHaydovchi\nusername: {d[1]}\nID:{id}\nBalans:{d[7]} ")
+            await bot.send_message(chat_id=d[3], text=f"Balansingiz {message.text} ga kamaytirildi")
+            await message.delete()
+            await state.finish()
+        else:
+            await message.answer("Kechirasiz bunday foydalanuvchi mavjud emas !")
+            await state.finish()
+
 @dp.callback_query_handler(text="banqilish")
 async def ban_user(call: types.CallbackQuery,state:FSMContext):
     await call.message.answer("Foydalanuvchi ID sini kiriting :")
@@ -158,15 +159,15 @@ async def id_get(message:Message,state:FSMContext):
         await message.answer("Iltimos son kiriting 🙅‍♂️🙅‍♂️🙅‍♂️")
     if message.text.isdigit():
         id = int(message.text)
-        try:
-
-            user = await db.select_user(id=id)
+        user = await db.select_user(telegram_id=id)
+        if user is not None:
             await state.update_data({"id":user[3]})
             user_of_banned.append(user[3])
             await message.answer("Necha kunga ban bo'lsin")
             await BanStatesGroup.day.set()
-        except TypeError:
+        else:
             await message.answer("Bu ID da foydalanuvchi mavjud emas . ")
+            await state.finish()
 @dp.message_handler(state=BanStatesGroup.day)
 async def how_many_day(message:Message,state:FSMContext):
     if message.text.isalpha():
@@ -179,7 +180,6 @@ async def how_many_day(message:Message,state:FSMContext):
             await message.answer(f"Foydalanuvchi {day} kun ichida bloklangan holatda bo'ladi")
             await message.delete()
             await state.finish()
-            # await asyncio.sleep(15)
             await asyncio.sleep(60*60*24*day)
             user_of_banned.remove(id)
     except TypeError:
@@ -199,7 +199,7 @@ async def id_get_ban_user(message:Message,state:FSMContext):
     if message.text.isdigit():
         id = int(message.text)
         try:
-            user = await db.select_user(id=id)
+            user = await db.select_user(telegram_id=id)
             if user[3] in user_of_banned:
                 user_of_banned.remove(user[3])
                 await message.answer("Foydalanuvchi blokdan chiqarildi")
